+++ exomiser-core/src/main/java/org/monarchinitiative/exomiser/core/model/frequency/FrequencyData.java
@@ -135,6 +135,41 @@ public class FrequencyData {
         return notNull;
     }
 
+    /**
+     *
+     * @return the count of the frequencies
+     * @since 13.3.0
+     */
+    public int size() {
+        return size;
+    }
+
+    /**
+     * Checks if there are any frequencies present. It is possible that there may be an rsID where there are no
+     * frequencies. For example, rare variants seen only in ClinVar.
+     *
+     * @return true if the count of the frequencies is zero
+     * @since 13.3.0
+     */
+    public boolean isEmpty() {
+        return size == 0;
+    }
+
+    /**
+     * Checks whether the instance contains a given {@link FrequencySource}.
+     * @param frequencySource The frequency source to check for membership in the instance.
+     * @return true if the instance contains the given {@link FrequencySource}
+     * @since 13.3.0
+     */
+    public boolean containsFrequencySource(FrequencySource frequencySource) {
+        for (FrequencySource dataSource : sources) {
+            if (dataSource == frequencySource) {
+                return true;
+            }
+        }
+        return false;
+    }
+
     public String getRsId() {
         return rsId;
     }
@@ -207,6 +242,24 @@ public class FrequencyData {
         return false;
     }
 
+    /**
+     * Returns the highest frequency in percent for a given set of {@link FrequencySource}.
+     *
+     * @param frequencySources to find the highest frequency for
+     * @return the population frequency as a percentage value i.e. 0.001 == 0.1%
+     * @since 13.3.3
+     */
+    public float getMaxFreqForPopulation(Set<FrequencySource> frequencySources) {
+        float max = 0f;
+        for (int i = 0; i < size; i++) {
+            FrequencySource freqSource = this.sources[i];
+            if (frequencySources.contains(freqSource)) {
+                max = Math.max(max, values[i]);
+            }
+        }
+        return max;
+    }
+
     @JsonIgnore
     public boolean hasKnownFrequency() {
         return size != 0;
@@ -246,7 +299,7 @@ public class FrequencyData {
     }
 
     /**
-     * Returns a the maximum frequency - if there are no known frequencies/ no
+     * Returns the maximum frequency - if there are no known frequencies/ no
      * frequency data it will return 0.
      *
      * @return