+++ exomiser-core/src/main/java/org/monarchinitiative/exomiser/core/writers/VcfResultsWriter.java
@@ -220,7 +220,7 @@ public class VcfResultsWriter implements ResultsWriter {
         fields.add(assignment.map(AcmgAssignment::acmgClassification).orElse(AcmgClassification.NOT_AVAILABLE).toString());
         fields.add(assignment.map(acmgAssignment -> toVcfAcmgInfo(acmgAssignment.acmgEvidence())).orElse(""));
         fields.add(assignment.map(acmgAssignment -> acmgAssignment.disease().getDiseaseId()).orElse(""));
-        fields.add('"' + assignment.map(acmgAssignment -> acmgAssignment.disease().getDiseaseName()).orElse("") + '"');
+        fields.add('"' + assignment.map(acmgAssignment -> acmgAssignment.disease().getDiseaseName().replace(" ", "_")).orElse("") + '"');
         return "{"+ String.join("|", fields) + "}";
     }