+++ exomiser-core/src/test/java/org/monarchinitiative/exomiser/core/genome/jannovar/JannovarDataSourceLoaderTest.java
@@ -20,12 +20,16 @@
 
 package org.monarchinitiative.exomiser.core.genome.jannovar;
 
+import de.charite.compbio.jannovar.UncheckedJannovarException;
 import de.charite.compbio.jannovar.data.JannovarData;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.CsvSource;
 
 import java.nio.file.Path;
 import java.nio.file.Paths;
 
+import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.CoreMatchers.instanceOf;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
@@ -36,17 +40,21 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 public class JannovarDataSourceLoaderTest {
 
     @Test
-    public void loadsOldFormatData() {
-        Path jannovarDataPath = Paths.get("src/test/resources/jannovar/1710_hg19_transcripts_ensembl.ser");
-        JannovarData jannovarData = JannovarDataSourceLoader.loadJannovarData(jannovarDataPath);
-        assertThat(jannovarData, instanceOf(JannovarData.class) );
+    public void failsToLoadPreviousNativeJannovarVersionFormatData() {
+        Path jannovarDataPath = Paths.get("src/test/resources/jannovar/1710_hg19_transcripts_ensembl_native_format.ser");
+        Exception exception = assertThrows(UncheckedJannovarException.class, () -> JannovarDataSourceLoader.loadJannovarData(jannovarDataPath));
+        assertThat(exception.getMessage(), equalTo("src/test/resources/jannovar/1710_hg19_transcripts_ensembl_native_format.ser was created by Jannovar 0.30 but we need at least 0.33-SNAPSHOT"));
     }
 
-    @Test
-    public void loadsNewFormatData() {
-        Path jannovarDataPath = Paths.get("src/test/resources/jannovar/1710_hg19_transcripts_ensembl_new_format.ser");
+    @ParameterizedTest
+    @CsvSource({
+            "src/test/resources/jannovar/2309_hg19_transcripts_ensembl_native_format.ser", // jannovar native format >= 0.33
+            "src/test/resources/jannovar/1710_hg19_transcripts_ensembl_exomiser_format.ser", // jannovar exomiser format < 0.33
+            "src/test/resources/jannovar/2309_hg19_transcripts_ensembl_exomiser_format.ser" // jannovar exomiser format >= 0.33
+    })
+    public void loadJannovarData(Path jannovarDataPath) {
         JannovarData jannovarData = JannovarDataSourceLoader.loadJannovarData(jannovarDataPath);
-        assertThat(jannovarData, instanceOf(JannovarData.class) );
+        assertThat(jannovarData, instanceOf(JannovarData.class));
     }
 
     @Test